refactor: consolidate KRAIL_WEBSITE_URL and deep link host into single constants
KRAIL_WEBSITE_URL ("https://krail.app") now lives in core/app-info alongside
appStoreUrl. KRAIL_DEEP_LINK_HOST/PATH ("ksharma-xyz.github.io", "/trip") live
in core/deeplink, shared by iOS and Android handlers. All inline literals and
the duplicate const in TripDeepLinkEncoder.kt are replaced with these imports.
