+++ core/deeplink/src/commonMain/kotlin/xyz/ksharma/krail/core/deeplink/DeepLinkConfig.kt
@@ -1,0 +1,4 @@
+package xyz.ksharma.krail.core.deeplink
+
+const val KRAIL_DEEP_LINK_HOST = "ksharma-xyz.github.io"
+const val KRAIL_DEEP_LINK_PATH = "/trip"
